+++ server/main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI, HTTPException, Request, Depends, Query
+from fastapi import FastAPI, HTTPException, Request, Depends
 from fastapi.responses import HTMLResponse
 from fastapi.security import OAuth2PasswordBearer
 from dotenv import load_dotenv
@@ -8,8 +8,8 @@ from datetime import datetime, timedelta, UTC
 from discord_webhook import DiscordWebhook, DiscordEmbed
 
 from db import UserRateDB
-from models import suggestGJStars20, Sort, Reassign, Auth, Clear, Rate
-from utils import AccountChecker
+from models import suggestGJStars20, Sort, Reassign, Auth, Clear, Rate, rateGJDemon21
+from utils import AccountChecker, featureToString
 
 load_dotenv()
 db = UserRateDB(f"mongodb+srv://{environ.get('MONGO_USERNAME')}:{environ.get('MONGO_PASSWORD')}@{environ.get('MONGO_ENDPOINT')}")
@@ -17,7 +17,6 @@ checker = AccountChecker()
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 SECRET_KEY = environ.get("SECRET_KEY")
-
 app = FastAPI(
     title="User Rate API",
     description="API for the Geode mod User Rate",
@@ -103,9 +102,6 @@ async def suggest(request: Request):
 
     form = suggestGJStars20(**form_dict)
 
-    if not form.gjp2:
-        return "-2"
-
     if form.stars is None or form.feature is None:
         return "-2"
 
@@ -128,7 +124,32 @@ async def suggest(request: Request):
 async def demon(request: Request):
     raw_form = await request.form()
     form_dict: dict = {key: value for key, value in raw_form.items()}
-    return {"message": "Demon"}
+
+    form_dict = convert_fields_to_int(form_dict, ["gameVersion", "binaryVersion", "accountID", "levelID", "rating"])
+    form_dict = convert_fields_to_bool(form_dict, ["gdw"])
+
+    if "rating" in form_dict and form_dict["rating"] not in [1, 2, 3, 4, 5]:
+        raise HTTPException(status_code=422, detail="Invalid demon rating")
+
+    form = rateGJDemon21(**form_dict)
+
+    if form.rating is None:
+        raise HTTPException(status_code=422, detail="Invalid demon rating")
+
+    sendDict = {
+        "ip": hashlib.sha256(request.client.host.encode()).hexdigest(),
+        "levelID": form.levelID,
+        "rating": form.rating.value,
+    }
+
+    level = db.rated_level(form.levelID)
+    if level is None or level["stars"] < 10:
+        return HTTPException(status_code=422, detail="Level must be rated demon")
+
+    db.demon(sendDict)
+    print(f"{request.client.host} suggested {form.rating} demon for level {form.levelID}")
+
+    return "-1"
 
 @app.get("/rated")
 async def check_rated_levels(levelIDs: str):
@@ -136,6 +157,10 @@ async def check_rated_levels(levelIDs: str):
         id_list = [int(id) for id in levelIDs.split(',')]
     except Exception:
         raise HTTPException(status_code=422, detail="Invalid level IDs")
+
+    if len(id_list) > 100:
+        raise HTTPException(status_code=422, detail="Too many level IDs")
+
     return db.check_rated_levels(id_list)
 
 @app.get("/mod/sent")
@@ -150,6 +175,7 @@ async def get_sent_levels(
         max_avg_feature: int = 4,
         token_data: dict = Depends(verify_token)
 ):
+    print(f"{token_data['accountID']} requested sent levels")
     return db.get_sent_levels(sort, page, min_send_count, max_send_count, min_avg_stars, max_avg_stars, min_avg_feature, max_avg_feature)
 
 @app.post("/mod/rate")
@@ -162,12 +188,14 @@ async def rate_level(data: Rate, token_data: dict = Depends(verify_token)):
     webhook = DiscordWebhook(url=environ.get("WEBHOOK_URL"), rate_limit_retry=True)
     embed = DiscordEmbed(
         title="Level Rated",
-        description=f"Level ID: {data.levelID}\nStars: {data.stars}\nFeature: {data.feature}",
+        description=f"Level ID: {data.levelID}\nStars: {data.stars}\nFeature: {featureToString(data.feature)}\nRated by: {token_data['accountID']}",
         color=0x00C04B
     )
     webhook.add_embed(embed)
 
     webhook.execute()
+
+    print(f"{token_data['accountID']} rated level {data.levelID} with {data.stars} stars and a {featureToString(data.feature)} rating")
 
     return {"message": "Success"}
 
@@ -181,6 +209,18 @@ async def reassign_moderator(data: Reassign, token_data: dict = Depends(verify_t
     else:
         db.demote_mod(data.accountID)
 
+    webhook = DiscordWebhook(url=environ.get("LOG_WEBHOOK_URL"), rate_limit_retry=True)
+    embed = DiscordEmbed(
+        title=f"Moderator {'Promoted' if data.promote else 'Demoted'}",
+        description=f"Account ID: {data.accountID}\nReassigned by: {token_data['accountID']}",
+        color=(0x00FF00 if data.promote else 0xFF0000)
+    )
+    webhook.add_embed(embed)
+
+    webhook.execute()
+
+    print(f"{token_data['accountID']} reassigned moderator {data.accountID} ({'promoted' if data.promote else 'demoted'})")
+
     return {"message": "Success"}
 
 @app.post("/admin/clear")
@@ -189,6 +229,40 @@ async def clear_sends(data: Clear, token_data: dict = Depends(verify_token)):
         raise HTTPException(status_code=400, detail="You must be an admin to clear sends")
 
     db.clear_sends_for_level(data.levelID)
+
+    webhook = DiscordWebhook(url=environ.get("LOG_WEBHOOK_URL"), rate_limit_retry=True)
+    embed = DiscordEmbed(
+        title="Sends Cleared",
+        description=f"Level ID: {data.levelID}\nCleared by: {token_data['accountID']}",
+        color=0xFF0000
+    )
+    webhook.add_embed(embed)
+
+    webhook.execute()
+
+    print(f"{token_data['accountID']} cleared sends for level {data.levelID}")
+
+    return {"message": "Success"}
+
+@app.post("/admin/derate")
+async def derate_level(data: Clear, token_data: dict = Depends(verify_token)):
+    if token_data["role"] != "admin":
+        raise HTTPException(status_code=400, detail="You must be an admin to derate levels")
+
+    db.derate(data.levelID)
+
+    webhook = DiscordWebhook(url=environ.get("LOG_WEBHOOK_URL"), rate_limit_retry=True)
+    embed = DiscordEmbed(
+        title="Level Derated",
+        description=f"Level ID: {data.levelID}\nDerated by: {token_data['accountID']}",
+        color=0xFF0000
+    )
+    webhook.add_embed(embed)
+
+    webhook.execute()
+
+    print(f"{token_data['accountID']} derated level {data.levelID}");
+
     return {"message": "Success"}
 
 @app.get("/mods")