+++ server/models.py
@@ -12,6 +12,13 @@ class Feature(Enum):
     LEGENDARY = 3
     MYTHIC = 4
 
+class Demon(Enum):
+    EASY = 1
+    MEDIUM = 2
+    HARD = 3
+    INSANE = 4
+    EXTREME = 5
+
 class Sort(IntEnum):
     TOP = 0
     RECENT = 1
@@ -31,6 +38,18 @@ class suggestGJStars20(BaseModel):
     levelID: int | None = None
     stars: int | None = None
     feature: Feature | None = None
+    gdw: bool = False
+    secret: str | None = None
+
+class rateGJDemon21(BaseModel):
+    gameVersion: int | None = None
+    binaryVersion: int | None = None
+    udid: str | None = None
+    uuid: str | None = None
+    accountID: int | None = None
+    gjp2: str | None = None
+    levelID: int | None = None
+    rating: Demon | None = None
     gdw: bool = False
     secret: str | None = None