+++ src/layers/SentLevelCell.hpp
@@ -13,10 +13,18 @@ class SentLevelCell final : public CCLayer {
 
     CCLabelBMFont* nameLabel = nullptr;
     DifficultyNode* diffNode = nullptr;
+    CCSprite* sendsSprite = nullptr;
+    CCLabelBMFont* sendsLabel = nullptr;
+    CCSprite* starsSprite = nullptr;
+    CCLabelBMFont* starsLabel = nullptr;
+    CCSprite* featureSprite = nullptr;
+    CCLabelBMFont* featureLabel = nullptr;
     CCMenu* buttonMenu = nullptr;
     CCMenuItemSpriteExtra* viewButton = nullptr;
 
     bool init(const SentLevel& level, float width, AdminPage* adminPage);
+
+    static std::string floatToString(float number, int precision);
 
     void onView(CCObject*);