+++ src/layers/ViewSentLevelPopup.hpp
@@ -8,21 +8,21 @@
 using namespace geode::prelude;
 
 class ViewSentLevelPopup final : public Popup<const SentLevel&, AdminPage*>{
-    // thanks B1rtek
-    // (https://github.com/B1rtek/Geode-GDDLIntegration/blob/8cecefc2199b7ade8a9266328893db075710f3b2/src/objects/RatingsSpread.h#L14)
     static constexpr int ratingColours[] = {
-        0xe67c73,
-        0xea8b71,
-        0xf2a96d,
-        0xf6b86b,
-        0xfac769,
-        0xffd666,
-        0xd5d06f,
-        0xabc978,
-        0x81c281,
-        0x57bb8a
+        0xffffff, // auto
+        0x00c8ff, // easy
+        0x00ff0a, // normal
+        0xffc300, // hard
+        0xffc300,
+        0xff4600, // harder
+        0xff4600,
+        0xff5de0, // insane
+        0xff5de0,
+        0x000000  // demon
     };
 
+    // thanks B1rtek
+    // (https://github.com/B1rtek/Geode-GDDLIntegration/blob/8cecefc2199b7ade8a9266328893db075710f3b2/src/objects/RatingsSpread.h#L14)
     static constexpr int featureColours[] = {
         0xe67c73,
         0xf2a96d,